Added the core module.
This implements a basic Queue and Cursor class.

The Cursor class is the old MySQLCursor and allows easy access to
the database by simply instancing. It uses thread-safe pooling of
connections.

The Queue class is the old manager.Queue and is now simpler to use
and less strict about getting exact times. The new queue implementation
also doesn't give priority to requests but instead makes both requests
and random songs equal. In turn there are less random songs in the queue
when there are no requests.